+++ test/libs/data-structures/IncrementalMerkleTree.test.ts
@@ -155,7 +155,11 @@ describe("IncrementalMerkleTree", () => {
 
         const directionBits = getDirectionBits(i, Number(await merkleTree.getUintTreeHeight()));
 
-        expect(await merkleTree.verifyUintProof(siblings, directionBits, elementHash)).to.be.true;
+        expect(await merkleTree.verifyUintProof(siblings, directionBits, elementHash, await merkleTree.getUintRoot()))
+          .to.be.true;
+        expect(await merkleTree.processIMTProof(siblings, directionBits, elementHash)).to.be.eq(
+          await merkleTree.getUintRoot(),
+        );
       }
     });
 
@@ -178,7 +182,8 @@ describe("IncrementalMerkleTree", () => {
 
       const directionBits = 0;
 
-      expect(await merkleTree.verifyUintProof(siblings, directionBits, elementHash)).to.be.true;
+      expect(await merkleTree.verifyUintProof(siblings, directionBits, elementHash, await merkleTree.getUintRoot())).to
+        .be.true;
     });
 
     it("should return false if proof is invalid", async () => {
@@ -194,7 +199,8 @@ describe("IncrementalMerkleTree", () => {
 
       const directionBits = 1;
 
-      expect(await merkleTree.verifyUintProof(siblings, directionBits, elementHash)).to.be.false;
+      expect(await merkleTree.verifyUintProof(siblings, directionBits, elementHash, await merkleTree.getUintRoot())).to
+        .be.false;
     });
 
     it("should return zeroHash if tree is empty", async () => {
@@ -336,7 +342,12 @@ describe("IncrementalMerkleTree", () => {
 
         const directionBits = getDirectionBits(i, Number(await merkleTree.getBytes32TreeHeight()));
 
-        expect(await merkleTree.verifyBytes32Proof(siblings, directionBits, elementHash)).to.be.true;
+        expect(
+          await merkleTree.verifyBytes32Proof(siblings, directionBits, elementHash, await merkleTree.getBytes32Root()),
+        ).to.be.true;
+        expect(await merkleTree.processIMTProof(siblings, directionBits, elementHash)).to.be.eq(
+          await merkleTree.getBytes32Root(),
+        );
       }
     });
   });
@@ -461,7 +472,12 @@ describe("IncrementalMerkleTree", () => {
 
         const directionBits = getDirectionBits(i, Number(await merkleTree.getAddressTreeHeight()));
 
-        expect(await merkleTree.verifyAddressProof(siblings, directionBits, elementHash)).to.be.true;
+        expect(
+          await merkleTree.verifyAddressProof(siblings, directionBits, elementHash, await merkleTree.getAddressRoot()),
+        ).to.be.true;
+        expect(await merkleTree.processIMTProof(siblings, directionBits, elementHash)).to.be.eq(
+          await merkleTree.getAddressRoot(),
+        );
       }
     });
   });