+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [patch]
+
+- Slightly extended Merkle Trees functionality. Added `processProof` function.
+
 ## [3.2.2]
 
 - Migrated to Hardhat 3.